requirements.txt
django
djangorestframework
psycopg2-binary
django_filter
django-rest-passwordreset
pyyaml
requests
python-dotenv
gunicorn
celery
redis
celery[redis]